+++ prototype_kl/parser/config.go
@@ -113,6 +113,9 @@ var SupportedFunctions = map[string]int{
 	"root":      2,
 	"sum":       2,
 	"avg":       -1, // variadic
+	"var":       -1, // variadic: population variance
+	"stddev":    -1, // variadic: population standard deviation
+	"median":    -1, // variadic: median of values
 	"fix":       2,
 	"prod":      2,
 	"min":       -1, // variadic